submodule, package renames
