--- titus-supplementary-component/es-tasks-publish-springboot/build.gradle
@@ -1,37 +1,0 @@
-/*
- * Copyright 2019 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-buildscript {
-    dependencies {
-        classpath "org.springframework.boot:spring-boot-gradle-plugin:2.1.2.RELEASE"
-    }
-}
-
-plugins {
-    id 'org.springframework.boot' version '2.1.2.RELEASE'
-}
-
-apply plugin: 'io.spring.dependency-management'
-
-dependencies {
-    compile project(':titus-supplementary-component-es-tasks-publish')
-
-    implementation 'org.springframework.boot:spring-boot-starter-web'
-    implementation 'org.springframework.boot:spring-boot-starter-tomcat'
-    implementation 'org.springframework.boot:spring-boot-starter-actuator'
-
-    compile project(':titus-testkit')
-}
