--- titus-supplementary-component/es-tasks-publish-springboot/src/main/resources/application.yml
@@ -1,4 +1,0 @@
-spring:
-  application:
-    name: titusestaskspublish
-
