--- titus-supplementary-component/es-tasks-publish/src/main/java/com/netflix/titus/supplementary/es/publish/TitusClient.java
@@ -1,29 +1,0 @@
-/*
- * Copyright 2019 Netflix, Inc.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.netflix.titus.supplementary.es.publish;
-
-import com.netflix.titus.grpc.protogen.Job;
-import com.netflix.titus.grpc.protogen.Task;
-import reactor.core.publisher.Flux;
-import reactor.core.publisher.Mono;
-
-public interface TitusClient {
-    Mono<Task> getTask(String taskId);
-
-    Flux<Task> getTaskUpdates();
-
-    Mono<Job> getJobById(String jobId);
-}
