+++ titus-supplementary-component/tasks-publisher-springboot/src/main/java/com/netflix/titus/supplementary/taskspublisher/TasksPublisherMain.java
@@ -1,0 +1,19 @@
+package com.netflix.titus.supplementary.taskspublisher;
+
+
+import com.netflix.titus.supplementary.taskspublisher.config.TasksPublisherConfiguration;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.context.annotation.Import;
+
+@SpringBootApplication
+@Import({TasksPublisherConfiguration.class})
+public class TasksPublisherMain {
+    private static final Logger logger = LoggerFactory.getLogger(TasksPublisherMain.class);
+
+    public static void main(String[] args) {
+        SpringApplication.run(TasksPublisherMain.class, args);
+    }
+}
